Move `maker_client.ts` & `taker_negotiator.ts` to `taker/`
