+++ src/index.ts
@@ -32,6 +32,6 @@ export {
   NetworkType
 } from "./negotiation/execution_params";
 export { MakerNegotiator } from "./negotiation/maker_negotiator";
-export { TakerNegotiator } from "./negotiation/taker_negotiator";
+export { TakerNegotiator } from "./negotiation/taker/taker_negotiator";
 
 export { TryParams } from "./timeout_promise";
